+++ examples/today-only/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @aldus-palace/example-today-only
 
+## 0.1.5
+
+### Patch Changes
+
+- Updated dependencies
+  - @aldus-palace/core@0.4.0
+
 ## 0.1.4
 
 ### Patch Changes